Deploy V3

+++ back/src/appointments/appointments.service.ts
@@ -8,6 +8,13 @@ import { UsersRepository } from 'src/users/users.repository';
 import { ServiceProfileRepository } from 'src/service-profile/service-profile.repository';
 import { AppointmentStatus } from 'src/enums/AppointmentStatus.enum';
 import { Appointment } from 'src/entities/appointment.entity';
+import { CreateAppointmentDto } from 'src/DTO/apptDtos/CreateAppointment.dto';
+import { User } from 'src/entities/user.entity';
+import { ServiceProfile } from 'src/entities/serviceProfile.entity';
+import { AppointmentToSaveDto } from 'src/DTO/apptDtos/appointmentToCreate.dto';
+import { UpdateAppointmentDto } from 'src/DTO/apptDtos/updateAppointment.dto';
+import { DeleteResult } from 'typeorm';
+import { ServiceProfileStatus } from 'src/enums/serviceProfileStatus.enum';
 
 @Injectable()
 export class AppointmentsService {
@@ -17,38 +24,80 @@ export class AppointmentsService {
     private readonly serviceProfileRepository: ServiceProfileRepository,
   ) {}
 
-  async createAppointmentService(createAppointment) {
-    const user = await this.usersRepository.getUserByIdRepository(
-      createAppointment.userId,
-    );
+  // CREAR UNA CITA
+  async createAppointmentService(
+    createAppointment: CreateAppointmentDto,
+    userId: string,
+  ): Promise<Appointment> {
+    const user: User = await this.usersRepository.getUserByIdRepository(userId);
     if (!user) throw new NotFoundException(`Usuario no encontrado`);
-    const provider =
+    const provider: ServiceProfile =
       await this.serviceProfileRepository.getServiceProfileByIdRepository(
         createAppointment.providerId,
       );
     if (!provider) throw new NotFoundException(`Proveedor no encontrado`);
-    return this.appointmentsRepository.createAppointmentRepository(
-      createAppointment,
-      user,
-      provider,
+
+    if (provider.status === ServiceProfileStatus.PENDING)
+      throw new BadRequestException(
+        `No se puede crear una cita con un perfil de servicio no activado`,
+      );
+
+    const appointmentToCreate: AppointmentToSaveDto = {
+      date: new Date(createAppointment.date),
+      additionalNotes: createAppointment.additionalNotes,
+      users: user,
+      provider: provider,
+    };
+
+    if (user.id === provider.user.id)
+      throw new BadRequestException(
+        `No se puede crear una cita con el perfil de servicio del mismo usuario`,
+      );
+
+    const savedAppointment: Appointment =
+      this.appointmentsRepository.createAppointmentRepository(
+        appointmentToCreate,
+      );
+
+    return await this.appointmentsRepository.saveAppointmentRepository(
+      savedAppointment,
     );
   }
 
-  async getAllAppointmentsService() {
+  // OBTENER TODAS LAS CITAS
+  async getAllAppointmentsService(): Promise<Appointment[]> {
     return this.appointmentsRepository.getAllAppointmentsRepository();
   }
 
-  async getAppointmentByIdService(id: string) {
-    return this.appointmentsRepository.getAppointmentByIdRepository(id);
+  // OBTENER UNA CITA BY ID
+  async getAppointmentByIdService(id: string): Promise<Appointment> {
+    const foundAppointment: Appointment =
+      await this.appointmentsRepository.getAppointmentByIdRepository(id);
+
+    if (!foundAppointment)
+      throw new NotFoundException(`No se pudo encontrar la cita con Id ${id}`);
+
+    return foundAppointment;
   }
 
-  async updateAppointmentByIdService(id: string, updateAppointment) {
-    const appointment =
+  // MODIFICAR UNA CITA BY ID
+  async updateAppointmentByIdService(
+    id: string,
+    updateAppointment: UpdateAppointmentDto,
+  ): Promise<Appointment> {
+    const appointment: Appointment =
       await this.appointmentsRepository.getAppointmentByIdRepository(id);
     if (!appointment)
-      throw new NotFoundException(`Appointment con Id ${id} no fue encontrado`);
-    const updated = Object.assign(appointment, updateAppointment);
-    return this.appointmentsRepository.updateAppointmentRepository(appointment);
+      throw new NotFoundException(`La cita con Id ${id} no fue encontrada`);
+
+    if (updateAppointment.date)
+      appointment.date = new Date(updateAppointment.date);
+    if (updateAppointment.additionalNotes)
+      appointment.additionalNotes = updateAppointment.additionalNotes;
+
+    return await this.appointmentsRepository.saveAppointmentRepository(
+      appointment,
+    );
   }
 
   // // CAMBIAR A CONFIRMADO
@@ -68,7 +117,7 @@ export class AppointmentsService {
 
   // CAMBIAR A COMPLETADO
   async completeAppointmentService(id: string): Promise<Appointment> {
-    const appointment =
+    const appointment: Appointment =
       await this.appointmentsRepository.getAppointmentByIdRepository(id);
 
     if (!appointment) {
@@ -82,12 +131,12 @@ export class AppointmentsService {
     }
 
     appointment.appointmentStatus = AppointmentStatus.COMPLETED;
-    return this.appointmentsRepository.updateAppointmentRepository(appointment);
+    return this.appointmentsRepository.saveAppointmentRepository(appointment);
   }
 
   // CAMBIAR A CANCELADO
   async cancelAppointmentService(id: string): Promise<Appointment> {
-    const appointment =
+    const appointment: Appointment =
       await this.appointmentsRepository.getAppointmentByIdRepository(id);
 
     if (!appointment) {
@@ -101,10 +150,20 @@ export class AppointmentsService {
     }
 
     appointment.appointmentStatus = AppointmentStatus.CANCELLED;
-    return this.appointmentsRepository.updateAppointmentRepository(appointment);
+    return this.appointmentsRepository.saveAppointmentRepository(appointment);
   }
 
-  deleteAppointmentService(id: string) {
-    return this.appointmentsRepository.deleteAppointmentRepository(id);
+  // ELIMINAR UNA CITA EN LA BASE DE DATOS
+  async deleteAppointmentService(id: string): Promise<Appointment> {
+    const foundAppointment: Appointment =
+      await this.appointmentsRepository.getAppointmentByIdRepository(id);
+    if (!foundAppointment)
+      throw new NotFoundException(`Appointmend con Id ${id} no fue encontrado`);
+    const result: DeleteResult =
+      await this.appointmentsRepository.deleteAppointmentRepository(id);
+    if (result.affected === 0) {
+      throw new NotFoundException('La cita no pudo ser eliminada.');
+    }
+    return foundAppointment;
   }
 }
